test mask generation

+++ image_understanding_prediction.py
@@ -83,7 +83,7 @@ for image_path in full_image_file_paths:
     # Transpose pred_mask to shape (H, W, N)
     if isinstance(pred_mask, torch.Tensor):
         pred_mask = pred_mask.cpu().numpy()
-        pred_mask = pred_mask > 0.5
+        pred_mask = (pred_mask > 0.5).astype(np.float32)
 
     N, H, W = pred_mask.shape
     pred_mask = pred_mask.transpose(1, 2, 0)  # Now pred_mask.shape == (H, W, N)
